src/feedback-fountain-api/requirements.txt
aiohttp==3.8.4
azure-ai-contentsafety==1.0.0b1
azure-cosmos==4.4.0
azure-identity==1.13.0
fastapi==0.95.2
mmh3==4.0.0
openai==0.27.7
python-dotenv==1.0.0
qdrant-client==1.1.7
tenacity==8.2.2
uvicorn==0.22.0
